requirements.txt
# NovaCura v2 — Full Requirements
# Core (already in v1)
flask>=2.3.0
python-dotenv>=1.0.1
flask-cors>=4.0.0
flask-socketio>=5.3.0
flask-limiter>=3.8.0
flask-login>=0.6.3
authlib>=1.3.1
PyJWT>=2.8.0
bcrypt>=4.1.3
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0
scipy>=1.11.0
plotly>=5.17.0
streamlit>=1.28.0
pandas>=2.0.0
shap>=0.43.0
reportlab>=4.0.0
xgboost>=2.0.3
lightgbm>=4.5.0
prometheus-client>=0.20.0
pip-audit>=2.7.3

# Upgrade 1: ChEMBL
requests>=2.31.0

# Upgrade 2: SMILES/RDKit
# Optional: rdkit-pypi has no wheels for some Python versions (for example 3.13 on Windows).
# Install manually when supported; backend includes a non-RDKit fallback for /predict-smiles.
# rdkit-pypi>=2022.9.5

# Upgrade 3: TA models (no extra deps)

# Upgrade 4: PostgreSQL
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
redis>=5.0.0

# Upgrade 5: Docker (no Python deps — docker-compose handles it)

# Upgrade 6: Active learning (no extra deps)

# Upgrade 7: LLM Analyst
anthropic>=0.25.0

# Upgrade 8: GNN
# Install separately (large packages):
# pip install torch --index-url https://download.pytorch.org/whl/cpu
# pip install torch_geometric

# Production
gunicorn>=21.2.0
eventlet>=0.33.0
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0
structlog>=23.0.0
